+++ projects/_02_sfc_internal/reconciliation/jobs/reconciliation.py
@@ -11,7 +11,7 @@ from utils import utils
 
 def main(
     cqc_locations_snapshot_source: str,
-    ascwds_reconciliation_source: str,
+    ascwds_workplace_source: str,
     reconciliation_single_and_subs_destination: str,
     reconciliation_parents_destination: str,
 ):
@@ -19,7 +19,7 @@ def main(
     spark.sql("set spark.sql.broadcastTimeout = 1000")
 
     cqc_location_df = utils.read_from_parquet(cqc_locations_snapshot_source)
-    ascwds_workplace_df = utils.read_from_parquet(ascwds_reconciliation_source)
+    ascwds_workplace_df = utils.read_from_parquet(ascwds_workplace_source)
 
     (
         first_of_most_recent_month,
@@ -72,7 +72,7 @@ if __name__ == "__main__":
 
     (
         cqc_locations_snapshot_source,
-        ascwds_reconciliation_source,
+        ascwds_workplace_source,
         reconciliation_single_and_subs_destination,
         reconciliation_parents_destination,
     ) = utils.collect_arguments(
@@ -81,8 +81,8 @@ if __name__ == "__main__":
             "Source s3 directory for latest run CQC locations snapshot dataset",
         ),
         (
-            "--ascwds_reconciliation_source",
-            "Source s3 directory for ASCWDS reconciliation parquet dataset",
+            "--ascwds_workplace_source",
+            "Source s3 directory for ASCWDS workplace parquet dataset",
         ),
         (
             "--reconciliation_single_and_subs_destination",
@@ -95,7 +95,7 @@ if __name__ == "__main__":
     )
     main(
         cqc_locations_snapshot_source,
-        ascwds_reconciliation_source,
+        ascwds_workplace_source,
         reconciliation_single_and_subs_destination,
         reconciliation_parents_destination,
     )